Refactored IO and Read classes + Graph only requires lists and map [Harman, Eric, Preet, Jainal]
