+++ src/main/java/fileio/IIO.java
@@ -9,14 +9,4 @@ import java.util.Map;
 
 public interface IIO {
 
-    void readFile();
-
-    Map<String, GraphNode> getNodeMap();
-
-    List<GraphEdge> getEdgeList();
-
-    //String lineRegex(String line, String regex);
-
-    //void writeFile(List<GraphNode> graphNodes, List<GraphEdge> graphEdge);
-
 }
